backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-bcrypt
Flask_testing
gunicorn
psycopg2-binary
Pytest


